+++ client/index.js
@@ -4,6 +4,7 @@ import i18next from "shared/translations/i18n";
 import ReactTooltip from "react-tooltip";
 import Notifications from "react-notify-toast";
 import Raven from "raven-js";
+import GoogleAnalytics from "common/components/GoogleAnalytics";
 
 import { Provider } from "react-redux";
 import { I18nextProvider } from "react-i18next";
@@ -13,14 +14,19 @@ import "./common/styles/entry.scss";
 import App from "./App";
 import store from "./common/store/store";
 
-// Load Sentry error reporting
 /* eslint-disable */
+
+// Load Google Analytics Tracking
+GoogleAnalytics.initialize();
+
+// Load Sentry error reporting
 if (SENTRY_ENABLED) {
 	Raven.config(SENTRY_DSN, {
 		release: BUILD_RELEASE,
 		environment: BUILD_ENVIRONMENT
 	}).install();
 }
+
 /* eslint-enable */
 
 ReactDOM.render(